refactor(website): move PartnerEngagementBand into the partners-marketplace shared layer
Sections may not import each other (enforced by check-conventions), so the
band shared by the pricing section and the customers pages belongs in the
shared composite layer, not under sections/.
