--- packages/twenty-website/src/sections/partner-engagement-band/index.ts
@@ -1,1 +1,0 @@
-export { PartnerEngagementBand } from './PartnerEngagementBand';
